💡 Update. import 경로 및 postgre에 맞게 JPQL 수정

+++ BE/rebirth/src/main/java/com/kkulmoo/rebirth/analysis/domain/dto/response/MonthlyLogInfoDTO.java
@@ -12,11 +12,11 @@ import lombok.*;
 public class MonthlyLogInfoDTO {
 
     @JsonProperty("total_spending_amount")
-    int totalSpendingAmount;
+    Long totalSpendingAmount;
 
     @JsonProperty("category_name")
     String categoryName;
 
     @JsonProperty("monthly_difference")
-    int monthlyDifferenceAmount;
+    Long monthlyDifferenceAmount;
 }
